requirements.txt
streamlit==1.31.0
pdfplumber==0.10.3
tabula-py==2.9.0
pytesseract==0.3.10
pdf2image==1.16.3
pandas==2.1.4
openpyxl==3.1.2
Pillow==10.2.0
PyPDF2==3.0.1
